ocr_service/requirements.txt
Click==7.0
Flask==1.0.2
flask-restplus==0.11.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
opencv-python==3.4.4.19
Pillow==5.3.0
Werkzeug==0.14.1
pytesseract==0.2.5
summa
PyPDF2
docxpy
